ClassCastException Resolved
An issued surroinding recieving the Complaint information from the
client and casting it to a Complaint within the Server was resolved by
adding the Entity Mappings to the Complaint class (Client side)  and
adding Hibernate files to the build path (Client side). Followed up by
changing out all instance of ServerModel.Complaint to
projectModel.Complaint.